# docs/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Historical Travel Agent API",
    "description": "歴史学習特化型旅行AIエージェント - REST API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/travel-plans": {
      "post": {
        "tags": [
          "travel-plans"
        ],
        "summary": "旅行計画を作成",
        "description": "旅行計画を作成する\n\nArgs:\n    request: 旅行計画作成リクエスト\n    repository: TravelPlanRepository\n\nReturns:\n    TravelPlanResponse: 作成された旅行計画\n\nRaises:\n    HTTPException: バリデーションエラー（400）",
        "operationId": "create_travel_plan_api_v1_travel_plans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTravelPlanRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "travel-plans"
        ],
        "summary": "旅行計画一覧を取得",
        "description": "ユーザーの旅行計画一覧を取得する\n\nArgs:\n    user_id: ユーザーID\n    repository: TravelPlanRepository\n\nReturns:\n    list[TravelPlanResponse]: 旅行計画リスト",
        "operationId": "list_travel_plans_api_v1_travel_plans_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TravelPlanResponse"
                  },
                  "title": "Response List Travel Plans Api V1 Travel Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/travel-plans/{plan_id}": {
      "get": {
        "tags": [
          "travel-plans"
        ],
        "summary": "旅行計画を取得",
        "description": "旅行計画を取得する\n\nArgs:\n    plan_id: 旅行計画ID\n    repository: TravelPlanRepository\n\nReturns:\n    TravelPlanResponse: 旅行計画\n\nRaises:\n    HTTPException: 旅行計画が見つからない（404）",
        "operationId": "get_travel_plan_api_v1_travel_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "travel-plans"
        ],
        "summary": "旅行計画を更新",
        "description": "旅行計画を更新する\n\nArgs:\n    plan_id: 旅行計画ID\n    request: 旅行計画更新リクエスト\n    repository: TravelPlanRepository\n\nReturns:\n    TravelPlanResponse: 更新された旅行計画\n\nRaises:\n    HTTPException: 旅行計画が見つからない（404）、バリデーションエラー（400）",
        "operationId": "update_travel_plan_api_v1_travel_plans__plan_id__put",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTravelPlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "travel-plans"
        ],
        "summary": "旅行計画を削除",
        "description": "旅行計画を削除する\n\nArgs:\n    plan_id: 旅行計画ID\n    repository: TravelPlanRepository\n\nReturns:\n    Response: 204 No Content\n\nRaises:\n    HTTPException: 旅行計画が見つからない（404）",
        "operationId": "delete_travel_plan_api_v1_travel_plans__plan_id__delete",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckpointSchema": {
        "properties": {
          "spotName": {
            "type": "string",
            "title": "Spotname"
          },
          "checkpoints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checkpoints"
          },
          "historicalContext": {
            "type": "string",
            "title": "Historicalcontext"
          }
        },
        "type": "object",
        "required": [
          "spotName",
          "checkpoints",
          "historicalContext"
        ],
        "title": "CheckpointSchema",
        "description": "チェックポイントスキーマ"
      },
...
      "CreateTravelPlanRequest": {
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1,
            "title": "Userid",
            "description": "ユーザーID"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "旅行タイトル"
          },
          "destination": {
            "type": "string",
            "minLength": 1,
            "title": "Destination",
            "description": "目的地"
          },
          "spots": {
            "items": {
              "$ref": "#/components/schemas/TouristSpotSchema"
            },
            "type": "array",
            "title": "Spots",
            "description": "観光スポットリスト"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "title",
          "destination"
        ],
        "title": "CreateTravelPlanRequest",
        "description": "旅行計画作成リクエスト"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoricalEventSchema": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "significance": {
            "type": "string",
            "title": "Significance"
          },
          "relatedSpots": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Relatedspots"
          }
        },
        "type": "object",
        "required": [
          "year",
          "event",
          "significance",
          "relatedSpots"
        ],
        "title": "HistoricalEventSchema",
        "description": "歴史イベントスキーマ"
      },
      "ImageAnalysisResponse": {
        "properties": {
          "detectedSpots": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detectedspots"
          },
          "historicalElements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Historicalelements"
          },
          "landmarks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Landmarks"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "detectedSpots",
          "historicalElements",
          "landmarks",
          "confidence"
        ],
        "title": "ImageAnalysisResponse",
        "description": "画像分析レスポンス"
      },
      "LocationSchema": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "description": "緯度"
          },
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "description": "経度"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "LocationSchema",
        "description": "地理的位置スキーマ"
      },
      "MapDataSchema": {
        "properties": {
          "center": {
            "$ref": "#/components/schemas/LocationSchema"
          },
          "zoom": {
            "type": "integer",
            "title": "Zoom"
          },
          "markers": {
            "items": {
              "$ref": "#/components/schemas/MapMarkerSchema"
            },
            "type": "array",
            "title": "Markers"
          }
        },
        "type": "object",
        "required": [
          "center",
          "zoom",
          "markers"
        ],
        "title": "MapDataSchema",
        "description": "地図データスキーマ"
      },
      "MapMarkerSchema": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng",
          "label"
        ],
        "title": "MapMarkerSchema",
        "description": "地図マーカースキーマ"
      },
...
      "ReflectionPhotoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "analysis": {
            "$ref": "#/components/schemas/ImageAnalysisResponse"
          },
          "userDescription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userdescription"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "analysis"
        ],
        "title": "ReflectionPhotoResponse",
        "description": "振り返り写真レスポンス"
      },
      "ReflectionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "planId": {
            "type": "string",
            "title": "Planid"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/ReflectionPhotoResponse"
            },
            "type": "array",
            "title": "Photos"
          },
          "userNotes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usernotes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "planId",
          "userId",
          "photos",
          "createdAt"
        ],
        "title": "ReflectionResponse",
        "description": "振り返りレスポンス"
      },
      "SpotDetailSchema": {
        "properties": {
          "spotName": {
            "type": "string",
            "title": "Spotname"
          },
          "historicalBackground": {
            "type": "string",
            "title": "Historicalbackground"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Highlights"
          },
          "recommendedVisitTime": {
            "type": "string",
            "title": "Recommendedvisittime"
          },
          "historicalSignificance": {
            "type": "string",
            "title": "Historicalsignificance"
          }
        },
        "type": "object",
        "required": [
          "spotName",
          "historicalBackground",
          "highlights",
          "recommendedVisitTime",
          "historicalSignificance"
        ],
        "title": "SpotDetailSchema",
        "description": "スポット詳細スキーマ"
      },
      "TouristSpotSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "スポットID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "スポット名"
          },
          "location": {
            "$ref": "#/components/schemas/LocationSchema",
            "description": "位置情報"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "説明"
          },
          "userNotes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usernotes",
            "description": "ユーザーメモ"
          }
        },
        "type": "object",
        "required": [
          "name",
          "location"
        ],
        "title": "TouristSpotSchema",
        "description": "観光スポットスキーマ"
      },
      "TravelGuideResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "planId": {
            "type": "string",
            "title": "Planid"
          },
          "overview": {
            "type": "string",
            "title": "Overview"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/HistoricalEventSchema"
            },
            "type": "array",
            "title": "Timeline"
          },
          "spotDetails": {
            "items": {
              "$ref": "#/components/schemas/SpotDetailSchema"
            },
            "type": "array",
            "title": "Spotdetails"
          },
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/CheckpointSchema"
            },
            "type": "array",
            "title": "Checkpoints"
          },
          "mapData": {
            "$ref": "#/components/schemas/MapDataSchema"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "planId",
          "overview",
          "timeline",
          "spotDetails",
          "checkpoints",
          "mapData",
          "createdAt",
          "updatedAt"
        ],
        "title": "TravelGuideResponse",
        "description": "旅行ガイドレスポンス"
      },
      "TravelPlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "spots": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Spots"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "guideGenerationStatus": {
            "type": "string",
            "title": "Guidegenerationstatus"
          },
          "reflectionGenerationStatus": {
            "type": "string",
            "title": "Reflectiongenerationstatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "guide": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TravelGuideResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "reflection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReflectionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "userId",
          "title",
          "destination",
          "spots",
          "status",
          "guideGenerationStatus",
          "reflectionGenerationStatus",
          "createdAt",
          "updatedAt"
        ],
        "title": "TravelPlanResponse",
        "description": "旅行計画レスポンス"
      },
      "UpdateTravelPlanRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "旅行タイトル"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination",
            "description": "目的地"
          },
          "spots": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TouristSpotSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spots",
            "description": "観光スポットリスト"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "旅行状態"
          }
        },
        "type": "object",
        "title": "UpdateTravelPlanRequest",
        "description": "旅行計画更新リクエスト"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}